Add a test case algorithm which always does the move
Refactor Metropolis to use MoveAlgorithm base class, which MoveAlways
also uses to avoid repetition
