+++ src/Metropolis.h
@@ -34,10 +34,11 @@
 // #include <CGAL/Mpzf.h>
 
 // CDT headers
-#include "Measurements.h"
-#include "MoveManager.h"
+//#include "Measurements.h"
+#include "MoveAlgorithm.h"
+//#include "MoveManager.h"
 #include "S3Action.h"
-#include "S3ErgodicMoves.h"
+//#include "S3ErgodicMoves.h"
 
 // C++ headers
 #include <algorithm>
@@ -50,18 +51,6 @@ using Gmpzf = CGAL::Gmpzf;
 
 extern const std::intmax_t PRECISION;
 
-/// @brief Convert enum class to its underlying type
-///
-/// http://stackoverflow.com/questions/14589417/can-an-enum-class-be-converted-to-the-underlying-type
-/// // NOLINT
-/// @tparam E Enum class type
-/// @param e Enum class
-/// @return Integral type of enum member
-template <typename E>
-constexpr auto to_integral(E e) -> typename std::underlying_type<E>::type
-{
-  return static_cast<typename std::underlying_type<E>::type>(e);
-}
 /// @class Metropolis
 /// @brief Metropolis-Hastings algorithm function object
 ///
@@ -71,12 +60,9 @@ constexpr auto to_integral(E e) -> typename std::underlying_type<E>::type
 /// \f[P_{ergodic move}=a_{1}a_{2}\f]
 /// \f[a_1=\frac{move[i]}{\sum\limits_{i}move[i]}\f]
 /// \f[a_2=e^{\Delta S}\f]
-class Metropolis
+class Metropolis : public MoveAlgorithm
 {
  private:
-  /// @brief A SimplicialManifold.
-  SimplicialManifold universe_;
-
   /// @brief The length of the timelike edges.
   long double Alpha_;
 
@@ -86,27 +72,6 @@ class Metropolis
   /// @brief \f$\lambda=\frac{\Lambda}{8\pi G_{N}}\f$ where \f$\Lambda\f$ is
   /// the cosmological constant.
   long double Lambda_;
-
-  /// @brief The current number of timelike edges
-  std::intmax_t N1_TL_{0};
-
-  /// @brief The current number of (3,1) and (1,3) simplices
-  std::intmax_t N3_31_13_{0};
-
-  /// @brief The current number of (2,2) simplices
-  std::intmax_t N3_22_{0};
-
-  /// @brief Number of passes of ergodic moves on triangulation.
-  std::intmax_t passes_{100};
-
-  /// @brief How often to print/write output.
-  std::intmax_t checkpoint_{10};
-
-  /// @brief Attempted (2,3), (3,2), (2,6), (6,2), and (4,4) moves.
-  Move_tracker attempted_moves_{};
-
-  /// @brief Successful (2,3), (3,2), (2,6), (6,2), and (4,4) moves.
-  std::array<std::atomic_intmax_t, 5> successful_moves_{};
 
  public:
   /// @brief Metropolis function object constructor
@@ -122,11 +87,7 @@ class Metropolis
   Metropolis(const long double Alpha, const long double K,
              const long double Lambda, const std::intmax_t passes,
              const std::intmax_t checkpoint)
-      : Alpha_(Alpha)
-      , K_(K)
-      , Lambda_(Lambda)
-      , passes_(passes)
-      , checkpoint_(checkpoint)
+      : MoveAlgorithm(passes, checkpoint), Alpha_(Alpha), K_(K), Lambda_(Lambda)
   {
 #ifndef NDEBUG
     std::cout << __PRETTY_FUNCTION__ << " called." << std::endl;
@@ -153,61 +114,6 @@ class Metropolis
   /// @return checkpoint_
   auto Checkpoint() const noexcept { return checkpoint_; }
 
-  /// @brief Gets attempted (2,3) moves.
-  /// @return attempted_moves_[0]
-  auto TwoThreeMoves() const noexcept { return attempted_moves_[0]; }
-
-  /// @brief Gets successful (2,3) moves.
-  /// @return successful_moves_[0]
-  auto SuccessfulTwoThreeMoves() const noexcept
-  {
-    return successful_moves_[0].load();
-  }
-
-  /// @brief Gets attempted (3,2) moves.
-  /// @return attempted_moves_[1]
-  auto ThreeTwoMoves() const noexcept { return attempted_moves_[1]; }
-
-  /// @brief Gets successful (3,2) moves.
-  /// @return std::get<1>(successful_moves_)
-  auto SuccessfulThreeTwoMoves() const noexcept
-  {
-    return successful_moves_[1].load();
-  }
-
-  /// @brief Gets attempted (2,6) moves.
-  /// @return return attempted_moves_[2]
-  auto TwoSixMoves() const noexcept { return attempted_moves_[2]; }
-
-  /// @brief Gets successful (2,6) moves.
-  /// @return std::get<2>(successful_moves_)
-  auto SuccessfulTwoSixMoves() const noexcept
-  {
-    return successful_moves_[2].load();
-  }
-
-  /// @brief Gets attempted (6,2) moves.
-  /// @return return attempted_moves_[3]
-  auto SixTwoMoves() const noexcept { return attempted_moves_[3]; }
-
-  /// @brief Gets successful (6,2) moves.
-  /// @return std::get<3>(attempted_moves_)
-  auto SuccessfulSixTwoMoves() const noexcept
-  {
-    return successful_moves_[3].load();
-  }
-
-  /// @brief Gets attempted (4,4) moves.
-  /// @return attempted_moves_[4]
-  auto FourFourMoves() const noexcept { return attempted_moves_[4]; }
-
-  /// @brief Gets successful (4,4) moves.
-  /// @return std::get<4>(attempted_moves_)
-  auto SuccessfulFourFourMoves() const noexcept
-  {
-    return successful_moves_[4].load();
-  }
-
   /// @brief Gets the total number of attempted moves.
   /// @return TwoThreeMoves() + ThreeTwoMoves() + TwoSixMoves() + SixTwoMoves()
   /// + FourFourMoves()
@@ -217,7 +123,7 @@ class Metropolis
            FourFourMoves();
   }
 
-  auto CurrentTotalSimplices() const noexcept { return N3_31_13_ + N3_22_; }
+  //  auto CurrentTotalSimplices() const noexcept { return N3_31_13_ + N3_22_; }
 
   /// @brief Calculate A1
   ///
@@ -343,126 +249,6 @@ class Metropolis
 
     return result;
   }  // CalculateA2()
-
-  void print_run()
-  {
-    std::cout << "Simplices: " << CurrentTotalSimplices() << std::endl;
-    std::cout << "Timeslices: "
-              << this->universe_.geometry->max_timevalue().get() << std::endl;
-    std::cout << "N3_31_13_: " << N3_31_13_ << std::endl;
-    std::cout << "N3_22_: " << N3_22_ << std::endl;
-    std::cout << "Timelike edges: " << N1_TL_ << std::endl;
-    std::cout << "Successful (2,3) moves: " << SuccessfulTwoThreeMoves()
-              << std::endl;
-    std::cout << "Attempted (2,3) moves: " << TwoThreeMoves() << std::endl;
-    std::cout << "Successful (3,2) moves: " << SuccessfulThreeTwoMoves()
-              << std::endl;
-    std::cout << "Attempted (3,2) moves: " << ThreeTwoMoves() << std::endl;
-    std::cout << "Successful (2,6) moves: " << SuccessfulTwoSixMoves()
-              << std::endl;
-    std::cout << "Attempted (2,6) moves: " << TwoSixMoves() << std::endl;
-    std::cout << "Successful (6,2) moves: " << SuccessfulSixTwoMoves()
-              << std::endl;
-    std::cout << "Attempted (6,2) moves: " << SixTwoMoves() << std::endl;
-    std::cout << "Successful (4,4) moves: " << SuccessfulFourFourMoves()
-              << std::endl;
-    std::cout << "Attempted (4,4) moves: " << FourFourMoves() << std::endl;
-  }
-
-  /// @brief Make a move of the selected type
-  ///
-  /// This function handles making a **move_type** move
-  /// by delegating to the particular named function, which handles
-  /// the bookkeeping for **attempted_moves_**. This function then
-  /// handles the bookkeeping for successful_moves_ and updates the
-  /// counters for N3_31_, N3_22_, and N1_TL_ accordingly.
-  ///
-  /// \done Add exception handling for moves to gracefully recover
-  /// \done Use MoveManager RAII class
-  ///
-  /// @param move The type of move
-  void make_move(const move_type move)
-  {
-#ifndef NDEBUG
-    std::cout << __PRETTY_FUNCTION__ << " called." << std::endl;
-#endif
-
-    // Make working copies
-    boost::optional<decltype(universe_)> maybe_moved_universe{universe_};
-    auto maybe_move_count = boost::make_optional(true, attempted_moves_);
-
-    // Initialize MoveManager
-    MoveManager<decltype(maybe_moved_universe), decltype(maybe_move_count)>
-        this_move(std::move(maybe_moved_universe), std::move(maybe_move_count));
-
-    // Setup moves
-    auto move_23_lambda =
-        [](SimplicialManifold manifold,
-           Move_tracker&      attempted_moves) -> SimplicialManifold {
-      return make_23_move(std::move(manifold), attempted_moves);
-    };
-    auto move_32_lambda =
-        [](SimplicialManifold manifold,
-           Move_tracker&      attempted_moves) -> SimplicialManifold {
-      return make_32_move(std::move(manifold), attempted_moves);
-    };
-    auto move_26_lambda =
-        [](SimplicialManifold manifold,
-           Move_tracker&      attempted_moves) -> SimplicialManifold {
-      return make_26_move(std::move(manifold), attempted_moves);
-    };
-    auto move_62_lambda =
-        [](SimplicialManifold manifold,
-           Move_tracker&      attempted_moves) -> SimplicialManifold {
-      return make_62_move(std::move(manifold), attempted_moves);
-    };
-
-    switch (move)
-    {
-      case move_type::TWO_THREE:
-      {
-        function_ref<SimplicialManifold(SimplicialManifold, Move_tracker&)>
-                                         move_function(move_23_lambda);
-        maybe_moved_universe = this_move.operator()(move_function);
-      }
-      break;
-      case move_type::THREE_TWO:
-      {
-        function_ref<SimplicialManifold(SimplicialManifold, Move_tracker&)>
-                                         move_function(move_32_lambda);
-        maybe_moved_universe = this_move.operator()(move_function);
-      }
-      break;
-      case move_type::TWO_SIX:
-      {
-        function_ref<SimplicialManifold(SimplicialManifold, Move_tracker&)>
-                                         move_function(move_26_lambda);
-        maybe_moved_universe = this_move.operator()(move_function);
-      }
-      break;
-      case move_type::SIX_TWO:
-      {
-        function_ref<SimplicialManifold(SimplicialManifold, Move_tracker&)>
-                                         move_function(move_62_lambda);
-        maybe_moved_universe = this_move.operator()(move_function);
-      }
-      break;
-      case move_type::FOUR_FOUR:
-        break;
-    }
-
-    // Check if move completed successfully and update if so
-    if (maybe_moved_universe) {
-      swap(universe_, maybe_moved_universe.get());
-      swap(attempted_moves_, this_move.attempted_moves_.get());
-      ++successful_moves_[to_integral(move)];
-    }
-
-    // Update counters
-    N1_TL_    = universe_.geometry->N1_TL();
-    N3_31_13_ = universe_.geometry->N3_31_13();
-    N3_22_    = universe_.geometry->N3_22();
-  }  // make_move()
 
   /// @brief Attempt a move of the selected type
   ///
